Moved symbol test to a better place Provided Symbol to dep package too
