+++ tools/mo/openvino/__init__.py
@@ -29,6 +29,7 @@ try:
     from openvino.runtime import InferRequest
     from openvino.runtime import AsyncInferQueue
 
+    from openvino.runtime import Symbol
     from openvino.runtime import Dimension
     from openvino.runtime import Strides
     from openvino.runtime import PartialShape
